// ui/src/lib/api/model/internalDeploymentsCreateDeploymentRunRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { InternalDeploymentsCreateDeploymentRunRequestConfig } from './internalDeploymentsCreateDeploymentRunRequestConfig';
import type { InternalDeploymentsCreateDeploymentRunRequestEnvironment } from './internalDeploymentsCreateDeploymentRunRequestEnvironment';

export interface InternalDeploymentsCreateDeploymentRunRequest {
  config?: InternalDeploymentsCreateDeploymentRunRequestConfig;
  environment?: InternalDeploymentsCreateDeploymentRunRequestEnvironment;
}